Modularise page layouts
ML experiment can be included/excluded by importing it in AgentNetwork
